Add setup::timers module
And move the Timers, SharedTimers, TimerToken and the two TimingWheel
aliases into it.

No functional changes, only moving code and changing the module
documentation (of setup and setup::timers).
