+++ rt/src/local.rs
@@ -15,7 +15,7 @@ use heph::actor_ref::{ActorGroup, ActorRef, SendError};
 use crate::info::Info;
 use crate::metrics::{LocalMetrics, SharedMetrics};
 use crate::scheduler::{self, Scheduler};
-use crate::setup::{TimerToken, Timers};
+use crate::setup::timers::{TimerToken, Timers};
 use crate::spawn::options::Priority;
 #[cfg(any(test, feature = "test"))]
 use crate::timing_wheel::TimingWheel;
