+++ rt/src/shared.rs
@@ -16,7 +16,7 @@ use crate::info::Info;
 use crate::metrics::SharedMetrics;
 use crate::scheduler::process::{FutureProcess, ProcessId};
 use crate::scheduler::shared::{Process, Scheduler};
-use crate::setup::{SharedTimers, TimerToken};
+use crate::setup::timers::{SharedTimers, TimerToken};
 #[cfg(test)]
 use crate::spawn::options::Priority;
 use crate::spawn::{ActorOptions, FutureOptions};
